Add additional asserts and ensure typing is correct

+++ HW Testing/Tests/utils.py
@@ -80,13 +80,13 @@ def ensureCalibrated(divecan_client: DiveCAN.DiveCAN,  shim_host: HWShim.HWShim)
     if message.data[1] == 0xFF and message.data[2] == 0xFF and message.data[3] == 0xFF:
         calibrateBoard(divecan_client, shim_host)
 
-def configureCell(shim_host: HWShim.HWShim, cellNum: int, cellType: configuration.CellType, cellVal:int):
+def configureCell(shim_host: HWShim.HWShim, cellNum: int, cellType: configuration.CellType, cellVal:float):
     if cellType == configuration.CellType.CELL_ANALOG:
         shim_host.set_analog_millis(cellNum, cellVal/2)
     elif cellType == configuration.CellType.CELL_DIGITAL:
         shim_host.set_digital_ppo2(cellNum, cellVal/100)
 
-def assertCell(cellType: configuration.CellType, cellVal:int, expectedCellVal: int):
+def assertCell(cellType: configuration.CellType, cellVal:float, expectedCellVal: float):
     if cellType == configuration.CellType.CELL_ANALOG: 
         # Check within 0.01 PPO2 or 1%
         # TODO: this spec is hot garbage and is likely a problem with the test stand, but need to be sure
